+++ R/africamonitor.R
@@ -9,6 +9,7 @@
 #' \code{\link[=am_sources]{am_sources()}}\cr
 #' \code{\link[=am_series]{am_series()}}\cr
 #' \code{\link{am_countries}}\cr
+#' \code{\link{am_countries_wld}}\cr
 #' \code{\link{am_entities}}
 #'
 #' Function to retrieve the data from the database
@@ -44,7 +45,7 @@
 #' @importFrom stats as.formula setNames
 #' @importFrom DBI dbConnect dbGetQuery dbDisconnect
 #' @importFrom RMySQL MySQL
-#' @importFrom collapse collapv ffirst fmedian funique settransformv get_vars get_vars<- date_vars add_vars add_vars<- cat_vars ss qF vlabels vlabels<- ckmatch qDT fnobs fnrow fncol unattrib namlab allNA
+#' @importFrom collapse collapv ffirst fmedian funique ftransformv get_vars get_vars<- date_vars add_vars add_vars<- cat_vars ss qF vlabels vlabels<- ckmatch qDT fnobs fnrow fncol unattrib namlab allNA
 #' @importFrom data.table setDT fifelse melt dcast transpose setcolorder
 # #' @importFrom writexl write_xlsx
 #'