+++ tests/testthat/test-accessdb.R
@@ -1,7 +1,7 @@
 
 data <- am_data(c("KEN", "NGA"), from = 2000)
 
-test_that("reshaping works", {
+if(length(data)) test_that("reshaping works", {
 
   expect_equal(data, am_pivot_wider(am_pivot_longer(data)))